assist: Split ai/model into smaller packages
