+++ lib/ai/model/output/parsejson.go
@@ -1,0 +1,30 @@
+package output
+
+import (
+	"encoding/json"
+	"strings"
+)
+
+// ParseJSONFromModel parses a JSON object from the model output and attempts to sanitize contaminant text
+// to avoid triggering self-correction due to some natural language being bundled with the JSON.
+// The output type is generic, and thus the structure of the expected JSON varies depending on T.
+func ParseJSONFromModel[T any](text string) (T, error) {
+	cleaned := strings.TrimSpace(text)
+	if strings.Contains(cleaned, "```json") {
+		cleaned = strings.Split(cleaned, "```json")[1]
+	}
+	if strings.Contains(cleaned, "```") {
+		cleaned = strings.Split(cleaned, "```")[0]
+	}
+	cleaned = strings.TrimPrefix(cleaned, "```json")
+	cleaned = strings.TrimPrefix(cleaned, "```")
+	cleaned = strings.TrimSuffix(cleaned, "```")
+	cleaned = strings.TrimSpace(cleaned)
+	var output T
+	err := json.Unmarshal([]byte(cleaned), &output)
+	if err != nil {
+		return output, newInvalidOutputErrorWithParseError(err)
+	}
+
+	return output, nil
+}
