+++ lib/ai/model/tools/embedding_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package model
+package tools
 
 import (
 	"context"
