+++ lib/ai/tokens/contants.go
@@ -1,0 +1,13 @@
+package tokens
+
+// Ref: https://github.com/openai/openai-cookbook/blob/594fc6c952425810e9ea5bd1a275c8ca5f32e8f9/examples/How_to_count_tokens_with_tiktoken.ipynb
+const (
+	// perMessage is the token "overhead" for each message
+	perMessage = 3
+
+	// perRequest is the number of tokens used up for each completion request
+	perRequest = 3
+
+	// perRole is the number of tokens used to encode a message's role
+	perRole = 1
+)
