+++ lib/ai/tokens/tokencount.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package model
+package tokens
 
 import (
 	"sync"
